+++ sha3/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update to `digest` v0.11
 - Replace type aliases with newtypes ([#678])
 - Implementation of the `SerializableState` trait ([#716])
+- Added default value for domain separator generic parameter on
+  `TurboShake128/256` types equal to 0x1F ([#798])
 
 ### Removed
 - `std` crate feature ([#678])
@@ -22,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#652]: https://github.com/RustCrypto/hashes/pull/652
 [#678]: https://github.com/RustCrypto/hashes/pull/678
 [#716]: https://github.com/RustCrypto/hashes/pull/716
+[#798]: https://github.com/RustCrypto/hashes/pull/798
 
 ## 0.10.8 (2023-04-08)
 ### Fixed